requirements.txt
reactpy==1.1.0
reactpy_router==2.0.0
pygame==2.6.1
openai==1.100.2
langchain==0.3.27
langchain-core==0.3.74
langchain-openai==0.3.30
langchain-text-splitters==0.3.9
python-dotenv==1.1.1
httpx==0.28.1
uvicorn[standard]
websockets==15.0.1
sounddevice
wavio
python-multipart
